Migrate `NodeInputsPage` to TS.
